Organize components. Update paper texture image with a wider and shorter image.
