+++ frontend/src/components/game/Controls.vue
