+++ frontend/src/components/misc/ApplePopup.vue
