+++ frontend/src/components/misc/CoffeeButton.vue
